+++ sca/npm/npmhandler.go
